Add files via upload
The Pharmacy Inventory Management System that we are going to implement will be covering all basic processes done in the pharmacy. So the staff does not need to spend time on writing stock records and updating them in files. And the serial number issuing process for medicines becomes easier and efficient. Keeping the track of stock and medical prescription details allow them to review the details whenever needed. 
The Pharmacy Inventory Management system is responsible for proper management of drug stocks, pop ups the notifications of expiry dates of stock items. This system allow the storekeeper to keep track of medicine stocks , notify the storekeeper when the stock is running out of medicines and help  him to reduce stock levels and eliminate stock waste.